Tests: Add render test for volume with at least one channel with zero extinction
A bug report was made about the odd rendering of volumes that have one
or more of it's colour channels set to zero. blender/blender#131723

This issue was fixed recently in 4.4 and this commit adds a render test
to make sure we don't accidentally break it again in the future.

Ref: blender/blender-test-data!34
Ref: blender/blender#123012

Pull Request: https://projects.blender.org/blender/blender/pulls/131960