Fix #89387: improve handling of unavailable links
A link is unavailable when any of its sockets are unavailable. Sometimes such
links are in a weird inbetween state between being there and not being there at
all. Generally, they should be considered to not be there, but sometimes it's
still useful if they automatically come back when toggling the availability of
sockets. We don't have a replacement for this functionality yet.

This patch changes behavior so that unavailable links are treated as not being
there at all in two places:
* When deciding whether to draw an input socket value.
* When deciding what to do when drawing from a socket (to create a link from it
  or detach existing links). This is done by simply removing unavailable links
  when starting to drag from a corresponding socket. This allows all of the
  existing logic to stay the same.

Pull Request: https://projects.blender.org/blender/blender/pulls/132184